+++ FiatChampAddon/FiatClient/AppConfig.cs
@@ -17,6 +17,7 @@ public record AppConfig
   [Required(AllowEmptyStrings = false)]
   public string SupervisorToken { get; set; } = null!;
   public FcaBrand Brand { get; set; }
+  public FcaRegion Region { get; set; }
   public string HomeAssistantUrl { get; set; } = "http://supervisor/core";
   public int StartDelaySeconds { get; set; } = 1; 
   public bool AutoRefreshLocation { get; set; } = false;